% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range.R
\name{as_date_range}
\alias{as_date_range}
\alias{date_range_query}
\alias{between_date_range}
\alias{check_date_range}
\title{Use lubridate to convert an object to a date range}
\usage{
as_date_range(
  x = NULL,
  year = NULL,
  days = 90,
  ...,
  start_date = NULL,
  end_date = NULL,
  limits = NULL,
  nm = c("start", "end"),
  call = caller_env()
)

date_range_query(x = NULL, .col = "date", ..., nm = c("start", "end"))

between_date_range(x = NULL, .col = "date", ..., nm = c("start", "end"))

check_date_range(
  x = NULL,
  ...,
  limits = NULL,
  nm = c("start", "end"),
  call = caller_env()
)
}
\arguments{
\item{x}{Date range as character vector in format of \code{c("<start date>", "<end date>")}. If length 1 and days is not \code{NULL}, return a range based on
\code{c(date_range, date_range + lubridate::days(days))}. \code{\link[=as_date_range]{as_date_range()}} also
allows a Date class object or a list of Date class objects. For all other
functions, x can also be a named list of Date objects with names matching
nm.}

\item{year}{If date_range is \code{NULL} and year is provided, date range is set
to \code{c("<year>-01-01", "<year>-12-31")}. year is ignored if date_range is
provided.}

\item{days}{Default range duration in days to use if date_range is length 1.}

\item{...}{
  Arguments passed on to \code{\link[lubridate:as_date]{lubridate::as_date}}
  \describe{
    \item{\code{tz}}{a time zone name (default: time zone of the POSIXt object \code{x}). See
\code{\link[=OlsonNames]{OlsonNames()}}.}
    \item{\code{origin}}{a Date object, or something which can be coerced by \code{as.Date(origin, ...)} to such an object (default: the Unix epoch of "1970-01-01"). Note that in
this instance, \code{x} is assumed to reflect the number of days since \code{origin} at
\code{"UTC"}.}
    \item{\code{format}}{format argument for character methods. When supplied parsing is
performed by \code{parse_date_time(x, orders = formats, exact = TRUE)}. Thus, multiple
formats are supported and are tried in turn.}
  }}

\item{start_date, end_date}{Start and end date used if year and date_range are
both \code{NULL}.}

\item{limits}{Optional range of allowed dates. If dates supplied to
\code{\link[=as_date_range]{as_date_range()}} falls outside limits range, abort function.}

\item{nm}{Names to use for returned date range list. Defaults to \code{c("start", "end")}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{.col}{Name of date column to use for query. Defaults to "date".}
}
\value{
A length 2 list with min and max Date values.
}
\description{
Use \code{\link[lubridate:as_date]{lubridate::as_date()}} to convert an object to a length 2 list with a
minimum and maximum date. By default the dates in the list will be named
named "start" and "end". \code{\link[=date_range_query]{date_range_query()}} is a variation that returns a
query string that can be passed to the "where" parameter of
\code{\link[=get_esri_data]{get_esri_data()}}. \code{\link[=between_date_range]{between_date_range()}} works the same way identical but
uses the BETWEEN DATE syntax. \code{\link[=check_date_range]{check_date_range()}} validates whether a date
or date range falls within supplied limits.
}
\examples{
as_date_range("2022-01-01", days = 10)

as_date_range(c("2022-01-01", "2022-01-31"))

as_date_range(year = 2022)

date_range_query(c("2022-01-01", "2022-01-31"))

# check_date_range("2022-09-01", limits = c("2022-07-01", "2022-09-30"))

}
