% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sys_num.R
\name{count_sys_num}
\alias{count_sys_num}
\title{Count the number of activated weapon system held by a country in each year}
\usage{
count_sys_num(data_frame, country_name)
}
\arguments{
\item{data_frame}{A dataframe from with the function extracts information}

\item{country_name}{The name of a country}
}
\value{
A dataframe that has been filtered and modified
}
\description{
Count the number of activated weapon system held by a country in each year
}
\examples{
count_sys_num(df_weapon_systems, 'RU')

}
