% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sf_data.R
\name{format_sf_data}
\alias{format_sf_data}
\alias{erase_data}
\title{Format simple feature data}
\usage{
format_sf_data(
  x,
  crs = getOption("getdata.crs", default = 3857),
  erase_data = NULL,
  dTolerance = NULL,
  smooth = FALSE,
  sf_col = NULL,
  sf_req = TRUE,
  ...
)

erase_data(x, erase_data = NULL)
}
\arguments{
\item{x}{A \code{sf} object or, if \code{sf_req} is \code{FALSE}, any object that can be
converted to an \code{sf} object with \link[sfext:as_sf]{sfext::as_sf}.}

\item{crs}{Coordinate reference system for returned data, Default: getOption("getdata.crs", default =
3857)}

\item{erase_data}{A \code{sf}, \code{sfc}, or \code{bbox} object with geometry that should be erased
from the data, Default: \code{NULL}}

\item{dTolerance}{numeric; tolerance parameter, specified for all or for each
feature geometry. If you run \code{st_simplify}, the input data is
specified with long-lat coordinates and \code{sf_use_s2()} returns
\code{TRUE}, then the value of \code{dTolerance} must be specified in
meters.}

\item{smooth}{If \code{TRUE}, smooth data with \link[smoothr:smooth]{smoothr::smooth} using default
method and parameters, Default: \code{FALSE}.}

\item{sf_col}{Name to use for output \code{sf} column, Default: 'geometry'.}

\item{sf_req}{If \code{TRUE}, data must be a \code{sf} object. If \code{FALSE}, data is
passed to \link[sfext:as_sf]{sfext::as_sf} to convert data to an \code{sf} object.}

\item{...}{Additional parameters passed to \code{format_data}}
}
\value{
A \code{sf} object with columns and geometry modified based parameters.
}
\description{
The main \link{format_sf_data} function is a wrapper for the following common
steps in transforming an \code{sf} object and preparing for mapping or analysis:
}
\details{
\itemize{
\item Convert data to an \code{sf} object with \link[sfext:as_sf]{sfext::as_sf} if \code{sf_req} is \code{FALSE}
\item Make data valid with \link[sf:st_make_valid]{sf::st_make_valid} if needed
\item Format data with \link{format_data} using the ... parameters
\item Erase any data overlapping with \code{erase_data} (suggested for use with water or open space)
\item Simplify geometry with \link[sf:st_simplify]{sf::st_simplify} if \code{dTolerance} is provided
\item Smooth geometry with \link[smoothr:smooth]{smoothr::smooth} if \code{smooth} is \code{TRUE}
\item Rename the sf column to match \code{sf_col} (defaults to "geometry")
}

The helper functions for format_sf_data and additional formatting functions
for \code{sf} data are described in the details.

Helper functions for \link{format_sf_data}:
\itemize{
\item \link{erase_data}: erase intersection of x and erase_data (validity of
erase_data checked before \link[sfext:st_erase]{sfext::st_erase} and for x after completing the
operation.
\item \link[sfext:rename_sf_col]{rename_sf_col}: Rename \code{sf} column.
\item \link[sfext:relocate_sf_col]{relocate_sf_col}: Relocate \code{sf} column after selected columns (defaults to
\code{\link[dplyr:everything]{dplyr::everything()}}).
}
}
\examples{
library(sf)

nc <- read_sf(system.file("shape/nc.shp", package = "sf"))
nc_county <- nc[2, ]

# Transform coordinate reference system
st_crs(nc)$epsg
st_crs(format_sf_data(nc, crs = 3857))$epsg

# Simplify and smooth geometry
plot(nc_county, max.plot = 1)
nc_county_simple <- format_sf_data(nc_county, dTolerance = 5000, smooth = TRUE)
plot(nc_county_simple, max.plot = 1)

# Erase data
nc_co_water <- get_tigris_data(
  type = "area water",
  state = "NC",
  county = nc_county$NAME
)
nc_county_erased <- format_sf_data(nc_county, erase_data = nc_co_water)
plot(nc_county_erased, max.plot = 1)

# If sf_req is set to FALSE, use any object that can be converted with sfext::as_sf
nc_bbox <- st_bbox(nc)
plot(format_sf_data(nc_bbox, erase_data = nc_county_simple, sf_req = FALSE))
}
