% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_data.R
\name{get_admin_data}
\alias{get_admin_data}
\alias{get_states}
\alias{get_counties}
\title{Get U.S. State and County boundary data (1:5 mi scale) for a location}
\usage{
get_states(
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  class = "df",
  call = caller_env(),
  ...
)

get_counties(
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  class = "df",
  ...
)
}
\arguments{
\item{location}{A \code{sf}, \code{sfc}, or \code{bbox} object or a character string that
matches a geoid, name, abb, or statefp for \link{us_states} or \link{us_counties}.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{class}{Class of data to return, "df" (default), "sf", "bbox", or "sfc"}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}

\item{...}{Additional parameters including geoid, name, state, or county.
These additional identifier parameters are only used if location is \code{NULL}.}
}
\description{
Get U.S. states and counties from packaged \link{us_counties} or \link{us_states} data.
See \code{\link[=get_tigris_data]{get_tigris_data()}} for more options.
}
\examples{
get_states(
  location = "Maryland"
)

get_counties(
  name = "Baltimore city, Maryland"
)

# Using short names with abbreviated state names for look-up may result in inexact matches
get_counties(
  county = "Baltimore, MD"
)

# Two-digit integer GeoIDs are supported
# bbox and wkt columns are dropped when returning class "sf"
get_states(
  geoid = 24,
  class = "sf"
)

# sf locations are used as a spatial filter
plot(
  get_counties(
    location = get_states("MD", class = "sf"),
    class = "sf"
  ),
  max.plot = 1
)
}
