% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_dir.R
\name{get_data_dir}
\alias{get_data_dir}
\alias{create_data_dir}
\title{Check data directory and create if needed}
\usage{
get_data_dir(path = NULL, package = "overedge")

create_data_dir(path = NULL, create = TRUE)
}
\arguments{
\item{path}{Path to directory to use as data directory.}

\item{package}{Package name; defaults to "overedge"}

\item{create}{For [create_data_dir], if `TRUE` and directory does not exist
at the path, prompt user to create directory. Defaults to `TRUE`.}
}
\description{
A utility function that wraps [rappdirs::user_cache_dir].
}
