% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_profile.R
\name{get_elev_profile}
\alias{get_elev_profile}
\title{Use \code{elevatr::get_elev_point} to get the elevation along a profile}
\usage{
get_elev_profile(
  profile,
  units = NULL,
  dist = FALSE,
  n = NULL,
  density = NULL,
  type = "regular",
  sample = NULL,
  ...,
  drop_units = FALSE,
  cumulative = FALSE
)
}
\arguments{
\item{profile}{A \code{sfc} or \code{sf} geometry with a LINESTRING or POINT geometry
type.}

\item{units}{If \code{NULL}, output elevation and distance is in meters. If a
valid distance unit is supplied, elevation and distance are converted to
match the supplied unit.}

\item{dist}{If \code{TRUE}}

\item{n}{integer; number of points to choose per geometry; if missing, n will be computed as \code{round(density * st_length(geom))}.}

\item{density}{numeric; density (points per distance unit) of the sampling, possibly a vector of length equal to the number of features (otherwise recycled); \code{density} may be of class \code{units}.}

\item{type}{character; indicate the sampling type, either "regular" or "random"}

\item{sample}{numeric; a vector of numbers between 0 and 1 indicating the points to sample - if defined sample overrules n, density and type.}

\item{...}{Unused at this time.}

\item{drop_units}{If \code{TRUE}, return a plain numeric column for the elevation
and distance columns. If \code{FALSE} (default) return a units class column.}

\item{cumulative}{If \code{TRUE}, and dist is \code{TRUE} return distance as a
cumulative sum.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{\link[=get_elev_profile]{get_elev_profile()}} is a wrapper for \code{\link[elevatr:get_elev_point]{elevatr::get_elev_point()}} that takes
a LINESTRING or POINT input for profile and returns a data frame of
elevation. Optionally, create a series of points along a line with
\code{\link[sf:st_line_sample]{sf::st_line_sample()}} and/or include a column with the distance between each
successive POINT in the data frame.
}
\details{
This function is proposed for addition to \code{{elevatr}}:
\url{https://github.com/elipousson/getdata/issues/4}
}
