% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esri_data.R
\name{get_esri_data}
\alias{get_esri_data}
\alias{get_esri_layers}
\alias{get_esri_metadata}
\title{Use esri2sf to get data from an ArcGIS FeatureServer or MapServer for a
location}
\usage{
get_esri_data(
  url,
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit"),
  asp = getOption("getdata.asp"),
  crs = getOption("getdata.crs", 3857),
  where = NULL,
  name = NULL,
  name_col = NULL,
  coords = NULL,
  from_crs = getOption("getdata.crs", 4326),
  clean_names = TRUE,
  token = NULL,
  progress = TRUE,
  quiet = FALSE,
  .name_repair = janitor::make_clean_names,
  ...,
  call = caller_env()
)

get_esri_layers(
  location = NULL,
  layers = NULL,
  url = NULL,
  nm = NULL,
  token = NULL,
  clean_names = TRUE,
  quiet = FALSE,
  .name_repair = janitor::make_clean_names,
  ...,
  call = caller_env()
)

get_esri_metadata(
  url,
  token = NULL,
  meta = NULL,
  clean_names = TRUE,
  .name_repair = janitor::make_clean_names,
  call = caller_env()
)
}
\arguments{
\item{url}{FeatureServer or MapServer url to retrieve data from. Passed to
\code{url} parameter of \code{\link[esri2sf:esri2sf]{esri2sf::esri2sf()}} or \code{\link[esri2sf:esri2df]{esri2sf::esri2df()}} functions.
For \code{\link[=get_esri_layers]{get_esri_layers()}}, the optional url must be a service url which is
the base url for one or more layer urls.}

\item{location}{\code{sf}, \code{sfc}, or \code{bbox} object (or other object convertible
with \code{\link[sfext:as_bbox]{as_bbox()}}. Optional.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Cordinate reference system to return, Default: 4326 for
\code{\link[sfext:sf_to_df]{sf_to_df()}} and \code{NULL} for \code{\link[sfext:df_to_sf]{df_to_sf()}}.}

\item{where}{where query string passed to esri2sf, Default: \code{NULL}}

\item{name, name_col}{Name value and name column found in the ArcGIS
FeatureServer or MapServer data.}

\item{coords}{Coordinate columns for input data.frame or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}

\item{from_crs}{For \code{\link[sfext:df_to_sf]{df_to_sf()}}, coordinate reference system used by
coordinates or well known text in data frame.}

\item{clean_names}{If \code{TRUE}, set .name_repair to
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}} Ignored when \code{\link[=get_esri_metadata]{get_esri_metadata()}} is not
returning a data.frame, e.g. \code{meta = "id"}.}

\item{token}{string for authentication token. defaults to \code{NULL}.}

\item{progress}{Show progress bar from \code{\link[cli:cli_progress_along]{cli::cli_progress_along()}} if \code{TRUE}.
Default \code{FALSE}.}

\item{quiet}{If \code{TRUE}, use \code{\link[=suppressMessages]{suppressMessages()}} to prevent the printing of
messages about the requested layer. Defaults to \code{FALSE}.}

\item{.name_repair}{Passed to \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.}

\item{...}{
  Arguments passed on to \code{\link[esri2sf:esri2sf]{esri2sf::esri2sf}}
  \describe{
    \item{\code{outFields}}{vector of fields you want to include. default is \code{NULL} for
all fields.}
    \item{\code{replaceDomainInfo}}{If \code{TRUE}, add domain information to the return data
frame. Default \code{FALSE}.}
    \item{\code{layerInfo}}{Layer information from \code{\link[esri2sf:esrimeta]{esrimeta()}}.}
    \item{\code{.fn}}{Function used to read items from a group layer. Defaults to
\code{\link[esri2sf:esri2sf]{esri2sf()}}.}
  }}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{layers}{Either a vector with URLs, a named list of urls, or a numeric
vector; defaults to \code{NULL}. Optional if url is a}

\item{nm}{Name or vector of names to add to the layers; defaults to \code{NULL}.}

\item{meta}{Name of metadata list value to return from \link[esri2sf:esrimeta]{esri2sf::esrimeta},
e.g. "name" to return layer name. Defaults to \code{NULL}.}
}
\description{
Wraps the \code{\link[esri2sf:esri2sf]{esri2sf::esri2sf()}} and \code{\link[esri2sf:esri2df]{esri2sf::esri2df()}} functions to download
an ArcGIS FeatureServer or MapServer. Supports spatial filtering with
bounding box based on location and filtering by location name (if location
name column is provided). As of fall 2022, this package suggests the
\href{https://github.com/elipousson/esri2sf/}{elipousson/esri2sf} fork using
httr2.
}
