% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsheet_data.R
\name{get_gsheet_data}
\alias{get_gsheet_data}
\title{Use googlesheets4 to get a data frame or simple feature data from a Google Sheet}
\usage{
get_gsheet_data(
  url,
  sheet = NULL,
  ss = NULL,
  ask = FALSE,
  geometry = FALSE,
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit", "meter"),
  asp = getOption("getdata.asp"),
  coords = getOption("getdata.coords", c("lon", "lat")),
  remove_coords = TRUE,
  address = getOption("getdata.address", "address"),
  geo = FALSE,
  from_crs = 4326,
  clean_names = TRUE,
  ...
)
}
\arguments{
\item{url}{A Google Sheets url}

\item{sheet}{Sheet to read, in the sense of "worksheet" or "tab". You can identify a sheet by name, with a string, or by position, with a number. Ignored if the sheet is specified via \code{range}. If neither argument specifies the sheet, defaults to the first visible sheet.}

\item{ss}{Something that identifies a Google Sheet:
\itemize{
\item its file id as a string or \code{\link[googledrive:drive_id]{drive_id}}
\item a URL from which we can recover the id
\item a one-row \code{\link[googledrive:dribble]{dribble}}, which is how googledrive
represents Drive files
\item an instance of \code{googlesheets4_spreadsheet}, which is what \code{\link[googlesheets4:gs4_get]{gs4_get()}}
returns
}

Processed through \code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}.}

\item{ask}{If \code{TRUE}, ask for the name of the Google Sheet to read if ss is
not provided to \link[sfext:read_sf_gsheet]{sfext::read_sf_gsheet}.}

\item{geometry}{Type of geometry to include in data frame. options include
"drop", "wkt", "centroid", "point", Default: 'centroid'.}

\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf:st_union]{sf::st_union()}}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{coords}{Coordinate columns for input data.frame or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}

\item{remove_coords}{For \code{\link[sfext:df_to_sf]{df_to_sf()}}, if \code{TRUE}, remove the coordinate columns
after converting a data frame to simple feature object; defaults to
\code{FALSE}.}

\item{address}{Address column name passed to \code{\link[tidygeocoder:geocode]{tidygeocoder::geocode()}} or
\link[tidygeocoder:geo]{tidygeocoder::geo}}

\item{geo}{If \code{TRUE}, use \code{\link[sfext:address_to_sf]{address_to_sf()}} to geocode address column; defaults
to \code{FALSE}.}

\item{from_crs}{For \code{\link[sfext:df_to_sf]{df_to_sf()}}, coordinate reference system used by
coordinates or well known text in data frame.}

\item{clean_names}{If \code{TRUE}, clean names provided to nm or created based on
value of col using \link[janitor:clean_names]{janitor::clean_names}. If \code{FALSE}, use names as
provided.}

\item{...}{Other parameters passed onto methods.}
}
\description{
Use googlesheets4 to get a data frame or simple feature data from a Google Sheet
}
