% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_data.R
\name{get_location_data}
\alias{get_location_data}
\alias{map_location_data}
\title{Get data for a location}
\usage{
get_location_data(
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit", default = "meter"),
  asp = getOption("getdata.asp"),
  data = NULL,
  pkg = getOption("getdata.package"),
  package = getOption("getdata.package"),
  fileext = getOption("getdata.fileext", default = "gpkg"),
  filetype = getOption("getdata.filetype", default = "gpkg"),
  fn = NULL,
  crop = TRUE,
  trim = FALSE,
  from_crs = getOption("getdata.from_crs"),
  crs = getOption("getdata.crs", 3857),
  class = "sf",
  label = NULL,
  index = NULL,
  col = NULL,
  var_names = NULL,
  clean_names = FALSE,
  range = NULL,
  .name_repair = "check_unique",
  ...,
  call = caller_env()
)

map_location_data(
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  data = NULL,
  package = NULL,
  fileext = "gpkg",
  filetype = "gpkg",
  fn = NULL,
  crop = TRUE,
  trim = FALSE,
  from_crs = NULL,
  crs = NULL,
  class = "list",
  label = NULL,
  load = FALSE,
  index = NULL,
  range = NULL,
  ...
)
}
\arguments{
\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf:st_union]{sf::st_union()}}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{data}{Character string (e.g. url, file path, or name of data from
package) for a spatial data or a \code{sf}, \code{sfc}, or \code{bbox}  object with
geometry overlapping the location. If data is \code{NULL}, all unnamed
parameters are passed to \code{\link[sfext:read_sf_ext]{sfext::read_sf_ext()}} with a bbox based on
location. If data is not \code{NULL} and not a data.frame, url, file path, or
bbox, conversion to a sf object will still always be attempted with
\code{\link[sfext:as_sf]{sfext::as_sf()}}.}

\item{pkg, package}{Name of the package to search for data.}

\item{fileext, filetype}{File extension or type to use if passing parameters
to \code{\link[sfext:read_sf_download]{sfext::read_sf_download()}} or \code{\link[sfext:read_sf_pkg]{sfext::read_sf_pkg()}} (required for
extdata and cached data).}

\item{fn}{Function to apply to data after filtering by location but before
returning from function.}

\item{crop}{If \code{TRUE}, x is cropped to y using \code{\link[sf:st_crop]{sf::st_crop()}}.}

\item{trim}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}

\item{from_crs}{Coordinate reference system used to match the location CRS to
the source data.}

\item{crs}{Coordinate reference system to return.}

\item{class}{Class of object to return.}

\item{label}{label is optionally used by \code{\link[=map_location_data]{map_location_data()}} to name the
data objects in the list returned by the function.}

\item{index}{A list of possible location, data, and (optionally) package
values. List must be named and include a value named package and package
must be \code{NULL}, to set package based on index. If list is not \code{NULL} and
location and/or data as character or numeric values, the location and data
are assumed to be index values for the index list. The index parameter
supports nested lists created by \code{\link[=make_location_data_list]{make_location_data_list()}} (using only
the default key names of "location" and "data"). This feature has not be
fully tested and may result in errors or unexpected results.}

\item{col}{For \link[sfext]{as_sf_list}, the name of the column used to group data if x
is a sf object or used to group and nest data before passing to x.}

\item{var_names}{A named list following the format, \code{list("New var name" = old_var_name)}, or a two column data frame with the first column being the
new variable names and the second column being the old variable names;
defaults to \code{NULL}.}

\item{clean_names}{If \code{TRUE}, clean names provided to nm or created based on
value of col using \link[janitor:clean_names]{janitor::clean_names}. If \code{FALSE}, use names as
provided.}

\item{range}{For \code{\link[sfext:lonlat_to_sfc]{lonlat_to_sfc()}}, an object that is coercible to a \code{bbox}
object or a length 4 vector with names xmin, xmax, ymin, and ymax. If a
coordinate pair falls outside the latitude/longitude range defined by the
vector but inside the range if reversed, the coordinates are assumed to be
in lat/lon order and are switched to lon/lat order before being converted
to a point. Defaults to \code{c("xmin" = -180, "ymin" = -50, "xmax" = 180, "ymax" = 60)}. Note that this default setting will reverse valid
coordinates north of Anchorage, Alaska or south of New Zealand.}

\item{.name_repair}{One of "unique", "universal", or "check_unique". See
\code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} for the meaning of these options.}

\item{...}{Additional parameters passed to \code{\link[sfext:read_sf_path]{sfext::read_sf_path()}},
\code{\link[sfext:read_sf_url]{sfext::read_sf_url()}}, \code{\link[sfext:read_sf_pkg]{sfext::read_sf_pkg()}}, \code{\link[sfext:as_sf]{sfext::as_sf()}} (with
bbox), or \code{\link[sfext:read_sf_ext]{sfext::read_sf_ext()}} (with no other parameters).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{load}{If \code{TRUE} and class is "list", load data to local environment;
defaults \code{FALSE}.}
}
\description{
Returns data for a selected location or a list of locations (for
\code{\link[=map_location_data]{map_location_data()}}). If data is a character string, the parameter is
passed to \code{\link[sfext:read_sf_url]{sfext::read_sf_url()}}, \code{\link[sfext:read_sf_path]{sfext::read_sf_path()}}, or
\code{\link[sfext:read_sf_pkg]{sfext::read_sf_pkg()}}. This function uses \code{\link[sfext:st_filter_ext]{sfext::st_filter_ext()}} to
filter, crop, or trim data to the provided location. location can also be an
an address.
}
\details{
This function previously supported county geoid, state name, abbreviation, or
geoid as a location. Currently, recommend using \code{\link[=get_tigris_data]{get_tigris_data()}} and
passing a \code{sf} object to location.

Working with sf lists for data and locations:

\code{\link[=map_location_data]{map_location_data()}} makes it easier to work with \code{sf} lists. It supports
data as a character vector, data as an \code{sf} list when location is a single
object, location as a character vector or \code{sf} list (including lists of
\code{bbox} or \code{sfc} objects), or when both data and location are lists (such as
a list created by \code{\link[=make_location_data_list]{make_location_data_list()}}).
}
