% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tigris_data.R
\name{get_tigris_data}
\alias{get_tigris_data}
\title{Use {tigris} to get state-level data from the U.S. Census Bureau}
\usage{
get_tigris_data(
  type = NULL,
  state = getOption("getdata.state"),
  name = NULL,
  crs = getOption("getdata.crs", default = 3857),
  cb = TRUE,
  clean_names = TRUE,
  ...
)
}
\arguments{
\item{type}{Type of data to return, Default: \code{NULL}; See details for
supported options.}

\item{state}{State name, abbreviation, or GeoID. Required. Defaults to
getOption("getdata.state").}

\item{name}{Name matching a value in the namelsad, namelsad, or geoid
columns. Default: \code{NULL}}

\item{clean_names}{If \code{TRUE}, pass data frame to \link[janitor:clean_names]{janitor::clean_names}; defaults to \code{TRUE}.}

\item{...}{Additional parameters passed on to {tigris} functions.}
}
\value{
A simple feature object matching the type provided.
}
\description{
Use the tigris package to download state-level data from the U.S. Census
Bureau API and filter by name or GeoID.
}
\details{
Supported data types:

Different type values corresponded to different tigris functions for
downloading from the U.S. Census Bureau API include. Supported options
include: "counties", "census places", "congressional districts",
"legislative districts", "senate district", "tracts", "block groups",
"blocks", "pumas", "voting districts", "zctas", "roads", "primary secondary
roads", "area water",  "linear water", and "landmarks".

tigris functions that do not use a "state" parameter (e.g.
\link[tigris:coastline]{tigris::coastline} or \link[tigris:rails]{tigris::rails})  are not supported by this
function.
}
