% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata-package.R
\docType{package}
\name{getdata-package}
\alias{getdata}
\alias{getdata-package}
\title{getdata: Get Easy Access to Tabular and Spatial Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Download and format spatial and non-spatial data with simple filtering by location.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/getdata}
  \item \url{https://elipousson.github.io/getdata/}
  \item Report bugs at \url{https://github.com/elipousson/getdata/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]
}

}
\keyword{internal}
