% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_location_data_list.R
\name{make_location_data_list}
\alias{make_location_data_list}
\title{Make a list of data and corresponding locations}
\usage{
make_location_data_list(data, location, key = c("location", "data"), ...)
}
\arguments{
\item{data, location}{A sf object or list of sf objects with data and
corresponding locations.}

\item{key}{Names for location and data in the returned list.}

\item{...}{Pass location_col and/or data_col to group and nest the data
provided to location and data. Use col to set both to the same value.}
}
\description{
This function converts data and location into lists of sf objects using
\link[sfext:as_sf_list]{as_sf_list}. If location_col, data_col, or col (which sets both to the
same value), are provided the col is passed to \link[sfext:as_sf_list]{as_sf_list} to allow the
creation of an sf list from a sf data frame using \code{\link[dplyr:group_nest]{dplyr::group_nest()}}.
}
\details{
If location and data are the same length are the same length, they are
combined into a single list. If either one is length 1 when the other is not,
the length 1 object is repeated to match the length of the longer object.
Different length objects where neither are length 1 gives a warning.
}
