% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_location_grid.R
\name{make_location_grid}
\alias{make_location_grid}
\title{Make a grid over the bounding box of a location}
\usage{
make_location_grid(location, name_col = "name", unit = NULL, ...)
}
\arguments{
\item{location}{A sf, sfc, or bbox object passed to \link[sfext:st_make_grid_ext]{sfext::st_make_grid_ext}}

\item{name_col}{Column name to collapse into new name_col value, Default:
'name'}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{...}{
  Arguments passed on to \code{\link[sfext:st_make_grid_ext]{sfext::st_make_grid_ext}}
  \describe{
    \item{\code{x}}{A \code{sf}, \code{sfc}, or \code{bbox} object, Default: \code{NULL}. Required.}
    \item{\code{crs}}{Coordinate reference system of bounding box to return; defaults to
\code{NULL} which maintains the crs of the input object.}
    \item{\code{ncol,nrow}}{Used to set n if either are not \code{NULL}; defaults to \code{NULL}.
row and id are added as columns to the grid if they are provided.}
    \item{\code{n}}{If n is NULL and square is \code{TRUE}, the grid is set automatically to
be 10 cells wide, Default: \code{NULL}}
    \item{\code{gutter}}{Distance in units between each column cell; gutter effectively
serves as a margin as the negative buffer is applied to all cells
(including those at the edges of the grid).}
    \item{\code{desc}}{If TRUE, reverse standard order of cell id numbering; defaults
\code{FALSE}}
    \item{\code{cellsize}}{numeric of length 1 or 2 with target cellsize: for square or rectangular cells the width and height, for hexagonal cells the distance between opposite edges (edge length is cellsize/sqrt(3)). A length units object can be passed, or an area unit object with area size of the square or hexagonal cell.}
    \item{\code{what}}{"polygons", "corners", "centers"; set to centers automatically if
style is "circle", "circle_offset" but a buffer is applied to return
circular polygons.}
    \item{\code{style}}{Style of cell to return with options including "rect", "square",
"hex", "flat_top_hex", "circle", "circle_offset"}
    \item{\code{.id}}{A name to use for the cell id column. Defaults to "id".}
    \item{\code{filter}}{If \code{TRUE} (or if trim is \code{TRUE}) filter grid geometry by x
using \link[sfext]{st_filter_ext}}
    \item{\code{trim}}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}
  }}
}
\description{
If location is a single feature sf object, the original columns of the object
are included in the output grid. If location has mutiple features, the values
of name_col are combined with \link[sfext:st_union_ext]{sfext::st_union_ext} and other columns are
dropped. The input sf object should not have columns named id, rows, or cols.
}
