% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{make_xwalk_list}
\alias{make_xwalk_list}
\title{Make a crosswalk list for use with \code{\link[=label_with_xwalk]{label_with_xwalk()}} or
\code{\link[=rename_with_xwalk]{rename_with_xwalk()}}}
\usage{
make_xwalk_list(xwalk, cols = c("label", "name"))
}
\arguments{
\item{cols}{Column names to use for crosswalk.}
}
\value{
A named list
}
\description{
Make a crosswalk list for use with \code{\link[=label_with_xwalk]{label_with_xwalk()}} or
\code{\link[=rename_with_xwalk]{rename_with_xwalk()}}
}
