% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osm_building_tags}
\alias{osm_building_tags}
\title{OpenStreetMap building tags}
\format{
A character vector with length of 84
}
\usage{
osm_building_tags
}
\description{
Used by \link{get_osm_data} if key = "building".
}
\details{
More information on the building key
\url{https://wiki.openstreetmap.org/wiki/Key:building}
}
\keyword{datasets}
