% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osm_common_tags}
\alias{osm_common_tags}
\title{Common OpenStreetMap tags}
\format{
A data frame with 272 rows and 5 variables:
\describe{
\item{\code{key}}{Key}
\item{\code{value}}{Value}
\item{\code{description}}{Description of tag/usage}
\item{\code{category}}{Category}
\item{\code{url}}{OSM Wiki url}
}
}
\usage{
osm_common_tags
}
\description{
A subset of tags scraped from the OpenStreetMap Wiki page on Map features:
\url{https://wiki.openstreetmap.org/wiki/Map_features}. Only those tags with a
url are included in this reference table.
}
\keyword{datasets}
