% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pkg_options.R
\name{set_pkg_options}
\alias{set_pkg_options}
\title{Set overedge packge options}
\usage{
set_pkg_options(
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  data_package = NULL,
  data_filetype = NULL,
  coords = NULL,
  from_crs = NULL,
  address = NULL,
  crs = NULL,
  state = NULL,
  overwrite = TRUE,
  pkg = "getdata"
)
}
\arguments{
\item{dist, diag_ratio, unit, asp, data_package, data_filetype, coords, from_crs, address, crs, state}{options to set, e.g. "crs = 2804" with \code{pkg = "getdata"} to set
"getdata.crs" to 2804.}

\item{overwrite}{If \code{TRUE}, overwrite any existing option value.}
}
\description{
Can set options for package, diag_ratio, dist, asp, or crs. Equivalent to
\code{\link[overedge:set_overedge_options]{overedge::set_overedge_options()}}
}
