% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{street_dir_prefixes}
\alias{street_dir_prefixes}
\title{Street directional prefixes}
\format{
A data frame with 8 rows and 3 variables:
\describe{
\item{\code{street_dir_abb}}{Street directional abbreviation}
\item{\code{street_dir_en}}{Street directional name (English)}
\item{\code{street_dir_es}}{Street directional name (Spanish)}
}
}
\usage{
street_dir_prefixes
}
\description{
A data frame based on Appendix I4 Directionals from the U.S. Postal Service
Publication 28 - Postal Addressing Standards. According to the U.S. Postal Service, "Directionals are not commonly used in
Puerto Rican addresses because other descriptions, such as the urbanization
name identify geographic areas. In the ZIP+4 file, the English equivalents
are used. Note: Although the Spanish word for West is Oeste, the abbreviation
W is used."
}
\details{
Source: \url{https://pe.usps.com/text/pub28/28api_007.htm}
}
\keyword{datasets}
