% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{street_suffixes}
\alias{street_suffixes}
\title{Street suffix abbreviations}
\format{
A data frame with 206 rows and 3 variables:
\describe{
\item{\code{street_suffix_abb}}{U.S. Postal Service standard suffix
abbreviation}
\item{\code{street_suffix}}{Street suffix name}
\item{\code{street_suffix_common}}{List column with commonly used street
suffix or abbreviation}
}
}
\usage{
street_suffixes
}
\description{
A data frame based on Appendix C1 Street Suffix Abbreviations from the U.S.
Postal Service Publication 28 - Postal Addressing Standards. This data
includes examples of suffix forms that are primary street suffix names,
common street suffixes or suffix abbreviations, and recommended official
U.S. Postal Service standard suffix abbreviations.
}
\details{
Source: \url{https://pe.usps.com/text/pub28/28apc_002.htm}
}
\keyword{datasets}
