% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_counties}
\alias{us_counties}
\title{U.S. County boundaries (1:5 mi scale, bbox and wkt)}
\format{
A data frame with 3,220 rows and 7 variables:
\describe{
  \item{`name`}{County name (tidycensus)}
  \item{`name_short`}{County name without state (tigris)}
  \item{`geoid`}{County GeoID (labelled)}
  \item{`abb_state`}{State abbreviation (USPS)}
  \item{`est_pop`}{Estimated county population (B01001_001), American Community Survey 5 year, 2015-2019}
  \item{`countyfp`}{County FIPS}
  \item{`statefp`}{State FIPS}
  \item{`bbox`}{Bounding box}
  \item{`wkt`}{Well known text (labelled)}
}
}
\usage{
us_counties
}
\description{
U.S. County boundaries data downloaded with [tigris::counties].
}
\details{
The geoid and wkt columns are labelled with a combination of an abbreviated
county name and state abbreviation in snake case. Population estimates
(est_pop) are not included for county equivalents in the U.S. Virgin Islands,
Guam, Northern Mariana Islands, and American Samoa.
}
\keyword{datasets}
