% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range.R
\name{as_date_range}
\alias{as_date_range}
\alias{date_range_query}
\title{Use lubridate to convert an object to a date range}
\usage{
as_date_range(
  x = NULL,
  year = NULL,
  days = 90,
  ...,
  start_date = NULL,
  end_date = NULL,
  nm = c("start", "end")
)

date_range_query(x = NULL, year = NULL, .col = "date", ...)
}
\arguments{
\item{x}{Date range as character vector in format of \code{c("<start date>", "<end date>")}. If length 1 and days is not \code{NULL}, return a range based on
\code{c(date_range, date_range + lubridate::days(days))}}

\item{year}{If date_range is NULL and year is provided, date range is set to
\code{c("<year>-01-01", "<year>-12-31")}. year is ignored if date_range is
provided.}

\item{days}{Default range duration in days to use if date_range is length 1.}

\item{...}{
  Arguments passed on to \code{\link[lubridate:as_date]{lubridate::as_date}}
  \describe{
    \item{\code{}}{}
  }}

\item{start_date, end_date}{Start and end date used if year and date_range are
both \code{NULL}.}

\item{nm}{Names to use for returned date range list. Defaults to c("start",
"end").}

\item{.col}{Name of date column to use for query. Defaults to "date".}
}
\value{
A length 2 list with min and max Date values.
}
\description{
Use \code{\link[lubridate:as_date]{lubridate::as_date()}} to convert an object to a length 2 list with a
minimum and maximum date. By default the dates in the list will be named
named "start" and "end". \code{\link[=date_range_query]{date_range_query()}} is a variation that returns a
query string that can be passed to the "where" parameter of \code{\link[=get_esri_data]{get_esri_data()}}.
}
\examples{
as_date_range("2022-01-01", days = 10)

as_date_range(c("2022-01-01", "2022-01-31"))

as_date_range(year = 2022)

date_range_query(c("2022-01-01", "2022-01-31"))

}
