% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_location_data.R
\name{cache_location_data}
\alias{cache_location_data}
\title{Cache location data with \code{sf::write_sf()} or \code{readr::write_rds()}}
\usage{
cache_location_data(
  data = NULL,
  ...,
  location = NULL,
  name = NULL,
  label = NULL,
  fileext = "gpkg",
  filename = NULL,
  path = NULL,
  prefix = NULL,
  postfix = NULL,
  cache = TRUE,
  pkg = "getdata",
  create = TRUE,
  overwrite = FALSE,
  call = caller_env()
)
}
\arguments{
\item{data}{Character string (e.g. url, file path, or name of data from
package) for a spatial data or a \code{sf}, \code{sfc}, or \code{bbox}  object with
geometry overlapping the location. If data is \code{NULL}, all unnamed
parameters are passed to \code{\link[sfext:read_sf_ext]{sfext::read_sf_ext()}} with a bbox based on
location. If data is not \code{NULL} and not a data.frame, url, file path, or
bbox, conversion to a sf object will still always be attempted with
\code{\link[sfext:as_sf]{sfext::as_sf()}}.}

\item{...}{
  Arguments passed on to \code{\link[=get_location_data]{get_location_data}}
  \describe{
    \item{\code{pkg,package}}{Name of the package to search for data.}
    \item{\code{fileext,filetype}}{File extension or type to use if passing parameters
to \code{\link[sfext:read_sf_ext]{sfext::read_sf_download()}} or \code{\link[sfext:read_sf_ext]{sfext::read_sf_pkg()}} (required for
extdata and cached data).}
    \item{\code{fn}}{Function to apply to data after filtering by location but before
returning from function.}
    \item{\code{from_crs}}{Coordinate reference system used to match the location CRS to
the source data.}
    \item{\code{crs}}{Coordinate reference system to return.}
    \item{\code{class}}{Class of object to return.}
    \item{\code{index}}{A list of possible location, data, and (optionally) package
values. List must be named and include a value named package and package
must be \code{NULL}, to set package based on index. If list is not \code{NULL} and
location and/or data as character or numeric values, the location and data
are assumed to be index values for the index list. The index parameter
supports nested lists created by \code{\link[=make_location_data_list]{make_location_data_list()}} (using only
the default key names of "location" and "data"). This feature has not be
fully tested and may result in errors or unexpected results.}
    \item{\code{dist}}{buffer distance in units. Optional.}
    \item{\code{diag_ratio}}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}
    \item{\code{asp}}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}
    \item{\code{unit}}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}
    \item{\code{crop}}{If \code{TRUE}, x is cropped to y using \code{\link[sf:st_crop]{sf::st_crop()}}.}
    \item{\code{trim}}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}
    \item{\code{col}}{For \link[sfext]{as_sf_list}, the name of the column used to group data if x
is a sf object or used to group and nest data before passing to x.}
    \item{\code{clean_names}}{If \code{TRUE}, clean names provided to nm or created based on
value of col using \link[janitor:clean_names]{janitor::clean_names}. If \code{FALSE}, use names as
provided.}
    \item{\code{var_names}}{A named list following the format, \code{list("New var name" = old_var_name)}, or a two column data frame with the first column being the
new variable names and the second column being the old variable names;
defaults to \code{NULL}.}
    \item{\code{range}}{For \code{\link[sfext:lonlat_to_sfc]{lonlat_to_sfc()}}, an object that is coercible to a \code{bbox}
object or a length 4 vector with names xmin, xmax, ymin, and ymax. If a
coordinate pair falls outside the latitude/longitude range defined by the
vector but inside the range if reversed, the coordinates are assumed to be
in lat/lon order and are switched to lon/lat order before being converted
to a point. Defaults to \code{c("xmin" = -180, "ymin" = -50, "xmax" = 180, "ymax" = 60)}. Note that this default setting will reverse valid
coordinates north of Anchorage, Alaska or south of New Zealand.}
  }}

\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \code{\link[sf:geos_combine]{sf::st_union()}}}

\item{name}{Name to make file name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}, e.g. "Residential zoning map" becomes
"residential_zoning_map". If the name includes a file extension it is
assumed that the filename has been provided as the name parameter.}

\item{label}{Label to combine with name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}. The label is designed to identify the area
or other shared characteristics across multiple data files, maps, or plots.
label is ignored if name is NULL or if name includes a file extension.}

\item{fileext}{File type or extension. Optional if filename or path include a
file extension.}

\item{filename}{File name; if filename is \code{NULL} and path does not include a
file extension, name and file extension are both required.}

\item{path}{Path to file or data directory. Optional. If path includes a file
extension and filename and fileext are both \code{NULL}, the filename and
extension included with path will be used instead. If multiple file
extensions are provided to filename, path, or fileext, \code{make_filename()}
will abort.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{postfix}{File name postfix; defaults to \code{NULL}.}

\item{cache}{If \code{TRUE}, path is set to the package cache directory using
\code{\link[filenamr:get_data_dir]{get_data_dir()}}; defaults to \code{FALSE}.}

\item{pkg}{Package name passed to appname parameter of
\code{\link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir()}}}

\item{create}{If \code{FALSE} and path does not exist, return path with a warning.
If \code{TRUE} and \code{\link[rlang:is_interactive]{rlang::is_interactive()}} is \code{TRUE}, ask user if directory
should be created. If the session not interactive and create is \code{TRUE}, a
new directory will be created.}

\item{overwrite}{If \code{TRUE}, remove a file with the same name and path}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Save data to file and invisibly return file path.
}
\description{
\code{\link[=cache_location_data]{cache_location_data()}} is a variant on \code{\link[=get_location_data]{get_location_data()}} that saves the
returned data to a cache directory using \code{\link[sf:st_write]{sf::write_sf()}} (if data is a sf
object) or \code{\link[readr:read_rds]{readr::write_rds()}} (if data is another class).
}
