% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_address_data.R
\name{format_address_data}
\alias{format_address_data}
\alias{bind_block_col}
\alias{bind_address_col}
\title{Format data frames and simple features with address data}
\usage{
bind_block_col(
  x,
  bldg_num = "bldg_num",
  street_dir_prefix = "street_dir_prefix",
  street_name = "street_name",
  street_suffix = "street_type",
  block_col = NULL,
  case = NULL
)

bind_address_col(x, city = NULL, county = NULL, state = NULL)
}
\arguments{
\item{bldg_num, street_dir_prefix, street_name, street_suffix}{Column names to
use for address information required to generate a block name and number.}

\item{block_col}{String to use as prefix for block identifier columns and
separator between block number and street. Set to "block" when \code{NULL}
(default).}

\item{case}{Case to use for new columns. "upper", "lower", or "title".
Defaults to \code{NULL}.}

\item{city, county, state}{City, county, and state to bind to data frame or
\code{sf} object.}
}
\description{
getdata has two helpers for working with address data:
\itemize{
\item \link{bind_address_col} bind a provided value for city, county, and state to a
data frame (to supplement address data with consistent values for these
variables)
\item \link{bind_block_col} requires a data frame with columns named "bldg_num",
"street_dir_prefix", "street_name", and "street_type" and binds derived
values for whether a building is on the even or odd side of a block and
create a block segment and a block face (including the even/odd identifier).
}
}
