% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\alias{rename_with_xwalk}
\alias{label_with_xwalk}
\alias{make_variable_dictionary}
\alias{fix_epoch_date}
\title{Format data frames and simple features using common approaches}
\usage{
format_data(
  x,
  var_names = NULL,
  xwalk = NULL,
  clean_names = TRUE,
  replace_na_with = NULL,
  replace_with_na = NULL,
  replace_empty_char_with_na = FALSE,
  fix_date = FALSE,
  label = FALSE,
  format_sf = FALSE,
  ...
)

rename_with_xwalk(
  x,
  xwalk = NULL,
  label = FALSE,
  .strict = TRUE,
  keep_all = TRUE,
  arg = caller_arg(x),
  call = caller_env()
)

label_with_xwalk(x, xwalk = NULL, label = "var", ...)

make_variable_dictionary(x, .labels = NULL, .definitions = NULL)

fix_epoch_date(x, .cols = dplyr::contains("date"))
}
\arguments{
\item{x}{A tibble or data frame object}

\item{var_names}{A named list following the format, \code{list("New var name" = old_var_name)}, or a two column data frame with the first column being the
new variable names and the second column being the old variable names;
defaults to \code{NULL}.}

\item{xwalk}{a data frame with two columns using the first column as name and
the second column as value; or a named list. The existing names of x must
be the values and the new names must be the names.}

\item{clean_names}{If \code{TRUE}, pass data frame to \link[janitor:clean_names]{janitor::clean_names};
defaults to \code{TRUE}.}

\item{replace_na_with}{A named list to pass to \link[tidyr:replace_na]{tidyr::replace_na}; defaults
to \code{NULL}.}

\item{replace_with_na}{A named list to pass to \link[naniar:replace_with_na]{naniar::replace_with_na};
defaults to \code{NULL}.}

\item{replace_empty_char_with_na}{If \code{TRUE}, replace "" with \code{NA} using
\link[naniar:replace_with_na_if]{naniar::replace_with_na_if}, Default: \code{TRUE}}

\item{fix_date}{If \code{FALSE}, fix UNIX epoch dates (common issue with dates from
FeatureServer and MapServer sources) using the \link{fix_epoch_date} function,
Default: \code{TRUE}}

\item{label}{For \code{\link[=label_with_xwalk]{label_with_xwalk()}} use \code{label = "val"} to use
\code{\link[labelled:val_labels]{labelled::set_value_labels()}} or "var" (default) to use
\code{\link[labelled:var_label]{labelled::set_variable_labels()}}. For \code{\link[=rename_with_xwalk]{rename_with_xwalk()}}, if label is
\code{TRUE}, xwalk is passed to \code{\link[=label_with_xwalk]{label_with_xwalk()}} with label = "var" to label
columns using the original names. Defaults to \code{FALSE}.}

\item{format_sf}{If \code{TRUE}, pass x and additional parameters to
\link{format_sf_data}.}

\item{...}{Additional parameters passed to \link{format_sf_data}}

\item{.strict}{If \code{TRUE} (default), require that all values from the xwalk
are found in the column names of the x data.frame. If \code{FALSE}, unmatched
values from the xwalk are ignored.}

\item{keep_all}{If \code{FALSE}, columns that are not named in the xwalk are
dropped. If \code{TRUE} (default), all columns are retained. If x is an \code{sf}
object, the geometry column will not be dropped even it is not renamed.}

\item{.labels}{Replaces labels column created by
\code{\link[labelled:look_for]{labelled::generate_dictionary()}} if column is all \code{NA} (no existing labels
assigned); defaults to \code{NULL}.}

\item{.definitions}{Character vector of definitions appended to dictionary
data frame. Must be in the same order as the variables in the provided data
frame x.}

\item{.cols}{tidyselect for columns to apply epoch date fixing function to.}
}
\value{
The input data frame or simple feature object with formatting
functions applied.
}
\description{
This function can apply the following common data cleaning tasks:
\itemize{
\item Applies \link[stringr:str_trim]{stringr::str_squish} and \link[stringr:str_trim]{stringr::str_trim} to all character
columns
\item Optionally replaces all character values of "" with \code{NA} values
\item Optionally corrects UNIX formatted dates with 1970-01-01 origins
\item Optionally renames variables by passing a named list of variables
}

The address functions previously included with \code{\link[=format_data]{format_data()}} are now
documented at \code{\link[=format_address_data]{format_address_data()}}.
}
