% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\alias{rename_with_xwalk}
\alias{label_with_xwalk}
\alias{fix_date}
\alias{bind_address_col}
\alias{bind_block_col}
\title{Format data frames and simple features using common approaches}
\usage{
format_data(
  x,
  var_names = NULL,
  clean_names = TRUE,
  replace_na_with = NULL,
  replace_with_na = NULL,
  replace_empty_char_with_na = TRUE,
  fix_date = FALSE,
  label = FALSE,
  format_sf = FALSE,
  xwalk = NULL,
  ...
)

rename_with_xwalk(x, xwalk = NULL, label = FALSE)

label_with_xwalk(x, xwalk)

fix_date(x)

bind_address_col(x, city = NULL, county = NULL, state = NULL)

bind_block_col(
  x,
  bldg_num = "bldg_num",
  street_dir_prefix = "street_dir_prefix",
  street_name = "street_name",
  street_suffix = "street_type"
)
}
\arguments{
\item{x}{A tibble or data frame object}

\item{var_names}{A named list following the format, \code{list("New var name" = old_var_name)}, or a two column data frame with the first column being the
new variable names and the second column being the old variable names;
defaults to \code{NULL}.}

\item{clean_names}{If \code{TRUE}, pass data frame to \link[janitor:clean_names]{janitor::clean_names}; defaults to \code{TRUE}.}

\item{replace_na_with}{A named list to pass to \link[tidyr:replace_na]{tidyr::replace_na}; defaults to
\code{NULL}.}

\item{replace_with_na}{A named list to pass to \link[naniar:replace_with_na]{naniar::replace_with_na};
defaults to \code{NULL}.}

\item{replace_empty_char_with_na}{If \code{TRUE}, replace "" with \code{NA} using
\link[naniar:replace_with_na_if]{naniar::replace_with_na_if}, Default: \code{TRUE}}

\item{fix_date}{If \code{TRUE}, fix UNIX dates (common issue with dates from
FeatureServer and MapServer sources) , Default: \code{TRUE}}

\item{label}{If \code{TRUE}, pass xwalk to \link{label_with_xwalk} to label columns
using the original names. Defaults to \code{FALSE}.}

\item{xwalk}{a data frame with two columns using the first column as name and
the second column as value; or a named list. The existing names of x must
be the values and the new names must be the names.}

\item{...}{Additional parameters passed to \link{format_sf_data}}

\item{city, county, state}{City, county, and state to bind to data frame or
\code{sf} object.}

\item{bldg_num, street_dir_prefix, street_name, street_suffix}{Column names to
use for address information required to generate a block name and number.}
}
\value{
The input data frame or simple feature object with formatting functions applied.
}
\description{
This function can apply the following common data cleaning tasks:
}
\details{
\itemize{
\item Applies \link[stringr:str_trim]{stringr::str_squish} and \link[stringr:str_trim]{stringr::str_trim} to all character
columns
\item Optionally replaces all character values of "" with \code{NA} values
\item Optionally corrects UNIX formatted dates with 1970-01-01 origins
\item Optionally renames variables by passing a named list of variables
}

Bind columns to address data:
\itemize{
\item \link{bind_address_col} bind a provided value for city, county, and state to a
data frame (to supplement address data with consistent values for these
variables)
\item \link{bind_block_col} requires a data frame with columns named "bldg_num",
"street_dir_prefix", "street_name", and "street_type" and binds derived
values for whether a building is on the even or odd side of a block and
create a block number (street segment), and block face (street segment side)
identifier.
}
}
