% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_airtable_data.R
\name{get_airtable_data}
\alias{get_airtable_data}
\title{Get Airtable data}
\usage{
get_airtable_data(
  base,
  table,
  record = NULL,
  fields = NULL,
  filter = NULL,
  sort = NULL,
  desc = FALSE,
  view = NULL,
  max_records = NULL,
  per_page = NULL,
  cell_format = "json",
  tz = NULL,
  locale = NULL,
  fields_by_id = FALSE,
  offset = NULL,
  token = NULL,
  list = "records",
  geometry = TRUE,
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit", "meter"),
  asp = getOption("getdata.asp"),
  crs = getOption("getdata.crs", 3857),
  coords = getOption("getdata.coords", c("lon", "lat")),
  from_crs = getOption("getdata.from_crs", 4326),
  remove_coords = TRUE,
  address = getOption("getdata.address", "address"),
  geo = FALSE,
  clean_names = TRUE,
  label = TRUE
)
}
\arguments{
\item{base}{Airtable base identifier}

\item{table}{Airtable table name or identifier}

\item{record}{Airtable record identifier, Default: NULL}

\item{fields}{Fields to return from Airtable base, Default: NULL}

\item{filter}{Filter (not currently supported), Default: NULL}

\item{sort}{Field to sort by, Default: NULL}

\item{desc}{If TRUE, sort in descending order, Default: FALSE}

\item{view}{Airtable view name(?) or identifier, Default: NULL}

\item{max_records}{Max records to return, Default: NULL}

\item{per_page}{Max records to return per page, Default: NULL}

\item{cell_format}{Cell format, Default: 'json'}

\item{tz}{Time zone, Default: NULL}

\item{locale}{Locale, Default: NULL}

\item{fields_by_id}{If TRUE, return fields by id, Default: FALSE}

\item{offset}{Offset parameter, Default: NULL}

\item{token}{API token, Default: NULL (same as get_access_token(type = "AIRTABLE_API_KEY"))}

\item{list}{Data type to return, Default: 'records'}

\item{geometry}{Type of geometry to include in data frame. options include
"drop", "wkt", "centroid", "point", Default: 'centroid'.}

\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \link[sf:geos_combine]{sf::st_union}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Cordinate reference system to return, Default: 4326 for \link[sfext]{sf_to_df}
and NULL for \link[sfext]{df_to_sf}.}

\item{coords}{Coordinate columns for input dataframe or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}

\item{from_crs}{For \link[sfext]{df_to_sf}, coordinate reference system used by
coordinates or well known text in data frame.}

\item{remove_coords}{For \link[sfext]{df_to_sf}, if \code{TRUE}, remove the coordinate columns
after converting a data frame to simple feature object; defaults to
\code{FALSE}.}

\item{address}{Address column name passed to \link[tidygeocoder:geocode]{tidygeocoder::geocode} or
\link[tidygeocoder:geo]{tidygeocoder::geo}}

\item{geo}{If \code{TRUE}, use \link[sfext]{address_to_sf} to geocode address column; defaults
to \code{FALSE}.}

\item{clean_names}{If \code{TRUE}, clean names provided to nm or created based on
value of col using \link[janitor:clean_names]{janitor::clean_names}. If \code{FALSE}, use names as
provided.}

\item{label}{label is optionally used by \link{map_location_data} to name the data
objects in the list returned by the function.}
}
\description{
Get Airtable data
}
