% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esri_data.R
\name{get_esri_data}
\alias{get_esri_data}
\alias{get_esri_layers}
\alias{get_esri_metadata}
\title{Use esri2sf to get data from an ArcGIS FeatureServer or MapServer for a
location}
\usage{
get_esri_data(
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit"),
  asp = getOption("getdata.asp"),
  crs = getOption("getdata.crs", 3857),
  url,
  where = NULL,
  name = NULL,
  name_col = NULL,
  coords = NULL,
  from_crs = getOption("getdata.crs", 4326),
  clean_names = TRUE,
  token = NULL,
  progress = TRUE,
  quiet = FALSE,
  ...
)

get_esri_layers(
  location = NULL,
  layers = NULL,
  url = NULL,
  nm = NULL,
  token = NULL,
  clean_names = TRUE,
  ...
)

get_esri_metadata(url, token = NULL, meta = NULL, clean_names = TRUE)
}
\arguments{
\item{location}{\code{sf}, \code{sfc}, or \code{bbox} object (or other object convertible
with \code{\link[=as_bbox]{as_bbox()}}. Optional.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Cordinate reference system to return, Default: 4326 for
\code{\link[sfext:sf_to_df]{sf_to_df()}} and NULL for \code{\link[sfext:df_to_sf]{df_to_sf()}}.}

\item{url}{FeatureServer or MapServer url to retrieve data from. Passed to
\code{url} parameter of \link[esri2sf:esri2sf]{esri2sf::esri2sf} or \link[esri2sf:esri2sf]{esri2sf::esri2df} functions. For
\link{get_esri_layers}, the optional url must be a service url which is the base
url for one or more layer urls.}

\item{where}{where query string passed to esri2sf, Default: \code{NULL}}

\item{name, name_col}{Name value and name column found in the ArcGIS
FeatureServer or MapServer data.}

\item{coords}{Coordinate columns for input dataframe or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}

\item{from_crs}{For \code{\link[sfext:df_to_sf]{df_to_sf()}}, coordinate reference system used by
coordinates or well known text in data frame.}

\item{clean_names}{If \code{TRUE}, use \code{\link[janitor:make_clean_names]{janitor::make_clean_names()}} on the
returned metadata value. Ignored for non-character values, e.g. \code{meta = "id"}.}

\item{token}{string for authentication token. defaults to \code{NULL}.}

\item{progress}{Show progress bar from \code{\link[cli:cli_progress_along]{cli::cli_progress_along()}} if \code{TRUE}.
Default \code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link[esri2sf:esri2sf]{esri2sf::esri2sf}}
  \describe{
    \item{\code{outFields}}{vector of fields you want to include. default is \code{NULL} for
all fields.}
    \item{\code{replaceDomainInfo}}{If \code{TRUE}, add domain information to the return data frame.
Default \code{FALSE}.}
  }}

\item{layers}{Either a vector with URLs, a named list of urls, or a numeric
vector; defaults to \code{NULL}. Optional if url is a}

\item{nm}{Name or vector of names to add to the layers; defaults to \code{NULL}.}

\item{meta}{Name of metadata list value to return from \link[esri2sf:esrimeta]{esri2sf::esrimeta},
e.g. "name" to return layer name. Defaults to \code{NULL}.}
}
\description{
Wraps the \link[esri2sf:esri2sf]{esri2sf::esri2sf} and \link[esri2sf:esri2sf]{esri2sf::esri2df} function to download an
ArcGIS FeatureServer or MapServer. Supports spatial filtering with bounding
box based on location and filtering by location name (if location name column
is provided). As of July 2022, this package suggests the
\href{https://github.com/elipousson/esri2sf/}{elipousson/esri2sf} fork using httr2
but I expect to see this fork merged back into the main package in the
future.
}
