% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esri_data.R
\name{get_esri_data}
\alias{get_esri_data}
\alias{get_esri_layers}
\alias{get_esri_metadata}
\title{Use esri2sf to get data from an ArcGIS FeatureServer or MapServer for a
location}
\usage{
get_esri_data(
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit"),
  asp = getOption("getdata.asp"),
  crs = getOption("getdata.crs", 3857),
  url,
  where = NULL,
  name = NULL,
  name_col = NULL,
  coords = NULL,
  from_crs = 4326,
  clean_names = TRUE,
  progress = TRUE,
  ...
)

get_esri_layers(location = NULL, layers, service_url = NULL, nm = NULL, ...)

get_esri_metadata(url, meta = "name", clean = TRUE)
}
\arguments{
\item{location}{\code{sf}, \code{sfc}, or \code{bbox} object (or other object convertible with
\code{\link[=as_bbox]{as_bbox()}}. Optional.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Cordinate reference system to return, Default: 4326 for \link[sfext]{sf_to_df}
and NULL for \link[sfext]{df_to_sf}.}

\item{url}{FeatureServer or MapServer url to retrieve data from. Passed to
\code{url} parameter of \link[esri2sf:esri2sf]{esri2sf::esri2sf} or
\link[esri2sf:esri2sf]{esri2sf::esri2df} functions.}

\item{where}{where query string passed to esri2sf, Default: \code{NULL}}

\item{name}{location name}

\item{name_col}{name of ArcGIS FeatureServer or MapServer column with
location names for features}

\item{coords}{Coordinate columns for input dataframe or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}

\item{from_crs}{For \link[sfext]{df_to_sf}, coordinate reference system used by
coordinates or well known text in data frame.}

\item{clean_names}{If \code{TRUE}, pass data frame to \link[janitor:clean_names]{janitor::clean_names}; defaults to \code{TRUE}.}

\item{progress}{Show progress bar from \code{\link[cli:cli_progress_along]{cli::cli_progress_along()}} if \code{TRUE}.
Default \code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link[esri2sf:esri2sf]{esri2sf::esri2sf}}
  \describe{
    \item{\code{outFields}}{vector of fields you want to include. default is \code{NULL} for
all fields.}
    \item{\code{geometry}}{An \code{sf} or \code{bbox} object. Currently, \code{sf} objects with a single
POINT feature are supported. All other \code{sf} objects are converted to \code{bbox}
objects.}
    \item{\code{bbox}}{bbox class object from \code{\link[sf:st_bbox]{sf::st_bbox()}} or a simple feature object
that can be converted to a bounding box.}
    \item{\code{token}}{string for authentication token. defaults to \code{NULL}.}
    \item{\code{geomType}}{string specifying the layer geometry ('esriGeometryPolygon'
or 'esriGeometryPoint' or 'esriGeometryPolyline' - if \code{NULL}, will try to
be inferred from the server)}
    \item{\code{spatialRel}}{Spatial relationship applied to the input \code{geometry} when
performing the query; defaults to \code{NULL} (equivalent to
"esriSpatialRelIntersects"). Additional supported options include
"esriSpatialRelContains", "esriSpatialRelCrosses",
"esriSpatialRelEnvelopeIntersects", "esriSpatialRelIndexIntersects",
"esriSpatialRelOverlaps", "esriSpatialRelTouches", "esriSpatialRelWithin"}
    \item{\code{replaceDomainInfo}}{If \code{TRUE}, add domain information to the return data frame.
Default \code{FALSE}.}
  }}

\item{layers}{Either a vector with URLs, a named list of urls, or a numeric
vector.}

\item{service_url}{Base service URL with layers are located.}

\item{nm}{Name or vector of names to add to the layers; defaults to \code{NULL}.}

\item{meta}{Name of metadata list value to return from \link[esri2sf:esrimeta]{esri2sf::esrimeta}.}

\item{clean}{If \code{TRUE}, use janitor::make_clean_names on the returned metadata
value (typically used for name values).}
}
\description{
Wraps the \link[esri2sf:esri2sf]{esri2sf::esri2sf} and \link[esri2sf:esri2sf]{esri2sf::esri2df} function to download an
ArcGIS FeatureServer or MapServer. Supports spatial filtering with bounding
box based on location and filtering by location name (if location name column
is provided).
}
\seealso{
\code{\link[esri2sf:esri2sf]{esri2sf::esri2sf()}}
}
