% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flickr_photos.R
\name{get_flickr_photos}
\alias{get_flickr_photos}
\title{Use FlickrAPI to get geotagged photos for a location}
\usage{
get_flickr_photos(
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  user_id = NULL,
  tags = NULL,
  license_id = "cc0",
  sort = "date-posted",
  desc = FALSE,
  img_size = "s",
  extras = c("description", "date_taken", "tags", "geo"),
  per_page = 100,
  page = 1,
  orientation = NULL,
  geometry = TRUE,
  crs = 4326,
  key = NULL
)
}
\arguments{
\item{location}{A `sf` or `bbox` object to use in creating bounding box for
getting photos from Flickr.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[overedge:get_asp]{get_asp()}} returns the same value
without modification.}

\item{user_id}{The NSID of the user who's photo to search. If this parameter
isn't passed then everybody's public photos will be searched.}

\item{tags}{A vector of tags you wish to search for.}

\item{license_id}{The license id for photos (for possible values see the
Flickr API method flickr.photos.licenses.getInfo). See details for more information.}

\item{sort}{Supported options include "date-posted","date-taken",
"interestingness", or "relevance"}

\item{desc}{If `TRUE` return images in descending sort order, if `FALSE`,
return in ascending sort order. Ignored if sort is set to "relevance".}

\item{img_size}{Defaults to "s" (small). Options ranging from smallest to
largest include "sq", "t", "s", "q", "m", "n", "z", "c", "l", and "o"}

\item{extras}{Defaults to "description", "date_taken", "tags", and "geo".}

\item{per_page}{Photos to return per page of search, Default: 100. Maximum
250 if location is provided or 500 otherwise.}

\item{page}{If page is greater than length 1, the function uses
`purrr::map_dfr()` to return results for all pages but this may cause
issues with API access if a large page range is provided. Default: 1}

\item{orientation}{If img_size is length 1, photos are filtered to one or
more of the supported orientations ("portrait", "landscape", and "square");
defaults to `NULL`.}

\item{geometry}{If `TRUE`, include "geo" in extras and convert photos data
frame to `sf` object.}

\item{crs}{Coordinate reference system of `sf` object to return if geometry
is `TRUE`.}
}
\value{
A data frame with photo information or `sf` object with geometry
  based on latitude and longitude of geocoded photos.
}
\description{
Set API key using `FlickrAPI::set_flickr_api_key()` or pass to the api_key
parameter. Currently relies on fork of package at
<https://github.com/elipousson/FlickrAPI>
}
\details{
License id options:

license_id can be an integer from 0 to 10 or a corresponding license code
including:

- "c" (All Rights Reserved),
- "by-bc-sa" (Attribution-NonCommercial-ShareAlike),
- "by-nc" (Attribution-NonCommercial),
- "by-nc-nd" (Attribution-NonCommercial-NoDerivs),
- "by" (Attribution),
- "by-sa" (Attribution-ShareAlike),
- "by-nd" (Attribution-NoDerivs),
- "nkc" (No known copyright restrictions),
- "pd-us" (United States Government Work),
- "cc0" (Public Domain Dedication),
- or "pd" (Public Domain Mark).
}
\seealso{
[FlickrAPI::getPhotoSearch()]
}
