% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_data.R
\name{get_location_data}
\alias{get_location_data}
\alias{map_location_data}
\title{Get data for a location}
\usage{
get_location_data(
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit", default = "meter"),
  asp = getOption("getdata.asp"),
  data = NULL,
  package = getOption("getdata.data_package"),
  filetype = getOption("getdata.data_filetype", default = "gpkg"),
  fn = NULL,
  crop = TRUE,
  trim = FALSE,
  from_crs = getOption("getdata.from_crs"),
  crs = getOption("getdata.crs"),
  class = "sf",
  label = NULL,
  index = NULL,
  col = NULL,
  var_names = NULL,
  clean_names = FALSE,
  ...
)

map_location_data(
  location = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  data = NULL,
  package = NULL,
  filetype = "gpkg",
  fn = NULL,
  crop = TRUE,
  trim = FALSE,
  from_crs = NULL,
  crs = NULL,
  class = "list",
  label = NULL,
  load = FALSE,
  index = NULL,
  ...
)
}
\arguments{
\item{location}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \link[sf:geos_combine]{sf::st_union}}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{data}{Character string (e.g. url, file path, or name of data from
package), a \code{sf}, \code{sfc}, or \code{bbox}  object including data in area.}

\item{package}{Name of the package to search for data.}

\item{filetype}{File type to use if passing parameters to \link[sfext:read_sf_ext]{sfext::read_sf_download}
or \link[sfext:read_sf_ext]{sfext::read_sf_pkg} (required for extdata and cached data).}

\item{fn}{Function to apply to data after filtering by location but before
returning from function.}

\item{crop}{If \code{TRUE}, x is cropped to y using \code{\link[sf:st_crop]{sf::st_crop()}}.}

\item{trim}{If \code{TRUE}, x is trimmed to y with \code{\link[sfext:st_trim]{st_trim()}}.}

\item{from_crs}{Coordinate reference system used to match the location CRS to
the source data.}

\item{crs}{Coordinate reference system to return.}

\item{class}{Class of object to return.}

\item{label}{label is optionally used by \link{map_location_data} to name the data
objects in the list returned by the function.}

\item{index}{A list of possible location, data, and (optionally) package
values. List must be named and include a value named package and package
must be \code{NULL}, to set package based on index. If list is not \code{NULL} and
location and/or data as character or numeric values, the location and data
are assumed to be index values for the index list. The index parameter
supports nested lists created by \link{make_location_data_list} (using only the
default key names of "location" and "data"). This feature has not be fully
tested and may result in errors or unexpected results.}

\item{col}{For \link[sfext]{as_sf_list}, the name of the column used to group data if x
is a sf object or used to group and nest data before passing to x.}

\item{var_names}{A named list following the format, \code{list("New var name" = old_var_name)}, or a two column data frame with the first column being the
new variable names and the second column being the old variable names;
defaults to \code{NULL}.}

\item{clean_names}{If \code{TRUE}, clean names provided to nm or created based on
value of col using \link[janitor:clean_names]{janitor::clean_names}. If \code{FALSE}, use names as
provided.}

\item{...}{additional parameters passed to \link[sfext:read_sf_ext]{sfext::read_sf_path}, \link[sfext:read_sf_ext]{sfext::read_sf_url}, or
\link[sfext:read_sf_ext]{sfext::read_sf_pkg} and \link[sfext:st_filter_ext]{sfext::st_filter_ext}}

\item{load}{If \code{TRUE} and class is "list", load data to local environment;
defaults \code{FALSE}.}
}
\description{
Returns data for a selected location or a list of locations (for
\link{map_location_data}). If data is a character string, the parameter is passed
to \link[sfext:read_sf_ext]{sfext::read_sf_url}, \link[sfext:read_sf_ext]{sfext::read_sf_path}, or \link[sfext:read_sf_ext]{sfext::read_sf_pkg}. This
function uses \code{\link[sfext:st_filter_ext]{sfext::st_filter_ext()}} to filter, crop, or trim data to the
provided location. location can also be an an address.
}
\details{
This function previously supported county geoid, state name, abbreviation, or
geoid as a location. Currently, recommend using \link{get_states} or
\link{get_counties} and passing a \code{sf} object to location.

Working with sf lists for data and locations:

\link{map_location_data} makes it easier to work with \code{sf} lists. It supports data
as a character vector, data as an \code{sf} list when location is a single object,
location as a character vector or \code{sf} list (including lists of \code{bbox} or \code{sfc}
objects), or when both data and location are lists (such as a list created
by \link{make_location_data_list}).
}
