% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tigris_data.R
\name{get_tigris_data}
\alias{get_tigris_data}
\title{Use tigris to get state-level data from the U.S. Census Bureau}
\usage{
get_tigris_data(
  type = NULL,
  state = getOption("getdata.state"),
  location = NULL,
  dist = getOption("getdata.dist"),
  diag_ratio = getOption("getdata.diag_ratio"),
  unit = getOption("getdata.unit", "meter"),
  asp = getOption("getdata.asp"),
  crs = getOption("getdata.crs", default = 3857),
  name = NULL,
  name_col = c("name", "namelsad", "geoid"),
  cb = TRUE,
  clean_names = TRUE,
  cache = TRUE,
  ...
)
}
\arguments{
\item{type}{Type of data to return, Default: \code{NULL}; See details for
supported options.}

\item{state}{State name, abbreviation, or GeoID. Required. Defaults to
getOption("getdata.state").}

\item{location}{A sf, sfc, or bbox object passed to \code{\link[sfext:st_bbox_ext]{sfext::st_bbox_ext()}}
and used to create the geometry passed to the \code{filter_by} parameter passed
to the tigris functions.}

\item{dist}{buffer distance in units. Optional.}

\item{diag_ratio}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}

\item{unit}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}

\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}

\item{crs}{Coordinate reference system of bounding box to return; defaults to
\code{NULL} which maintains the crs of the input object.}

\item{name, name_col}{Name and columns to filter by name. name defaults to
\code{NULL}, and name_col defaults to c("namelsad", "namelsad", "geoid")
columns.}

\item{cb}{If \code{TRUE}, download a generalized (1:500k) file. If \code{FALSE},
download the most detailed TIGER/Line file. Defaults to \code{TRUE} (reverse of
the default for tigris functions). This parameter is \emph{not} used when type
is set to blocks, roads, primary secondary roads, area water, linear water,
landmarks, or zctas.}

\item{clean_names}{If \code{TRUE}, set .name_repair to \code{\link[janitor:make_clean_names]{janitor::make_clean_names()}};
defaults to \code{TRUE}.}

\item{cache}{If \code{TRUE}, set \code{options(tigris_use_cache = TRUE)} to cache
downloaded tigris data. Ignored if \code{getOption("tigris_use_cache")} is not
\code{NULL}.}

\item{...}{Additional parameters passed on to tigris functions.}
}
\value{
A simple feature object matching the type provided.
}
\description{
Use the \href{https://github.com/walkerke/tigris}{{tigris}} package to download
state-level data from the U.S. Census Bureau API and optionally filter by
name or GeoID.
}
\details{
Supported data types:

Different type values corresponded to different tigris functions for
downloading from the U.S. Census Bureau API include. Supported options
include: "counties", "census places", "congressional districts",
"legislative districts", "senate district", "tracts", "block groups",
"blocks", "pumas", "voting districts", "zctas", "roads", "primary secondary
roads", "area water",  "linear water", and "landmarks".

tigris functions that do not use a "state" parameter (e.g.
\link[tigris:coastline]{tigris::coastline} or \link[tigris:rails]{tigris::rails})  are not supported by this
function. Note that the default value of the cb parameter for
\link{get_tigris_data} is \code{TRUE} and the default value of for the original
{tigris} package is \code{FALSE}.
}
