% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata-package.R
\docType{package}
\name{getdata-package}
\alias{getdata}
\alias{getdata-package}
\title{getdata: Get Easy Access to Tabular and Spatial Data}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
