% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_sf_ext}
\alias{st_bbox_ext}
\alias{st_filter_ext}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sfext}{\code{\link[sfext]{read_sf_ext}}, \code{\link[sfext]{st_bbox_ext}}, \code{\link[sfext]{st_filter_ext}}}
}}

