% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_address_data.R
\name{replace_with_xwalk}
\alias{replace_with_xwalk}
\alias{replace_street_suffixes}
\alias{replace_street_dir_prefixes}
\title{Replace values in a character vector or data frame with a crosswalk}
\usage{
replace_with_xwalk(
  x,
  .cols = NULL,
  xwalk = NULL,
  dict = NULL,
  abb = TRUE,
  case = NULL
)

replace_street_suffixes(
  x,
  street_suffix = "street_type",
  xwalk = NULL,
  case = NULL,
  abb = TRUE
)

replace_street_dir_prefixes(
  x,
  street_dir_prefix = "street_dir_prefix",
  xwalk = NULL,
  case = NULL,
  abb = TRUE
)
}
\arguments{
\item{.cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to transform.
Because \code{across()} is used within functions like \code{summarise()} and
\code{mutate()}, you can't select or compute upon grouping variables.}

\item{xwalk, dict}{Named list or data frame with a minimum of two columns
where one column contains the replacement values and the other the values
to replace. If xwalk is \code{NULL}, dict is used and vice-versa. If both are
provided, the xwalk values take precedence so they can be used to override
a dict or add new values.}

\item{abb}{If abb is \code{TRUE} (default), the second column of the dict is
assumed to be abbreviation that should be used as the replace for the
values in x or the replacement column. Otherwise, the first column is
assumed to hold the replacement values and the second column is assumed to
hold the original values. For example, for \code{\link[=replace_street_suffixes]{replace_street_suffixes()}}, If
\code{TRUE}, replace full suffix names with abbreviations. If \code{FALSE}, replace
abbreviations with full street suffix names.}
}
\description{
Replace values in a character vector or data frame with a crosswalk
}
