% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_access_token.R
\name{set_access_token}
\alias{set_access_token}
\alias{set_token_type}
\alias{get_access_token}
\alias{get_token_type}
\title{Set or get an access token or API key to/from environment variables.}
\usage{
set_access_token(
  token,
  overwrite = FALSE,
  install = FALSE,
  type = NULL,
  call = caller_env(),
  .frame = parent.frame()
)

get_access_token(token = NULL, type = NULL, call = caller_env())
}
\arguments{
\item{token}{An access token or API key; required for \code{\link[=set_access_token]{set_access_token()}}.
If token is not provided; type is required for \code{\link[=get_access_token]{get_access_token()}}. Length
1 character vector or list. If named, the name of the token is used in
place of type.}

\item{overwrite}{If \code{TRUE}, overwrite existing token; Default: \code{FALSE}}

\item{install}{If \code{TRUE}, install token for use in future sessions; Default:
\code{FALSE}}

\item{type}{Name of token; defaults to \code{NULL}. Optional if token is named.}

\item{call}{Passed as the call parameter for \code{\link[cli:cli_abort]{cli::cli_abort()}} to improve
error messages when function is used internally.}
}
\description{
Based on the \code{\link[mapboxapi:mb_access_token]{mapboxapi::mb_access_token()}} function from the {mapboxapi}
package by Kyle Walker.
}
