% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_access_token.R
\name{set_access_token}
\alias{set_access_token}
\alias{get_access_token}
\title{Set or get an access token or API key to/from environment variables.}
\usage{
set_access_token(token, overwrite = FALSE, install = FALSE, type = NULL)

get_access_token(token = NULL, type = NULL)
}
\arguments{
\item{token}{An access token or API key; required for [set_access_token()].
If token is not provided; type is required for [get_access_token()].}

\item{overwrite}{If `TRUE`, overwrite existing token; Default: `FALSE`}

\item{install}{If `TRUE`, install token for use in future sessions; Default:
`FALSE`}

\item{type}{Name of token; defaults to `NULL`.}
}
\description{
Based on the [mapboxapi::mb_access_token] function from the {mapboxapi}
package by Kyle Walker.
}
